Remove std::vector where possible.
Many caluclations based around equations of a specific degree
will only result in a known maximum number of solutions,
meaning that dynamic allocation is not necessary and a static,
'vector-like' data structure can be used instead.

The StaticVector class was created to be a drop-in replacement
for std::vector, but with a set maximum size and no dynamic
allocation. This should be a major performance improvement.